Renamed SettingsLoader to PersistentSettings
